Make excess recursion fatal
This is a bit of a Heisenbug, because running with debug on makes
Verilator generate many errors and quit on reaching the error count
limit, while with debug off it outputs the first two of them and then
gets stuck in an infinite loop allocating more and more memory.

So we just stop processing on such an error to avoid surprises.