Improve performance of cont assignments with delay
If there already exists a sentree that could be used for an always we
create from an `AssignW`, we use it. This way no new trigger is created,
and we get the benefit of likely reducing the number of `eval`s. That is
because then the always is only resumed if any input changes, as opposed
to possibly multiple times per clock period.

Signed-off-by: Krzysztof Bieganski <kbieganski@antmicro.com>